+++ src/collective/nitf/testing.py
@@ -3,12 +3,13 @@
 
 We install collective.cover to test the availibility and features of
 the tile included for that package.
+
+For Plone 5 we need to install plone.app.contenttypes.
 """
 from PIL import Image
 from plone.app.robotframework.testing import AUTOLOGIN_LIBRARY_FIXTURE
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
-from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import PloneSandboxLayer
 from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
@@ -19,8 +20,17 @@ from zope.component import getUtility
 from zope.intid.interfaces import IIntIds
 
 import os
+import pkg_resources
 import random
 import string
+
+
+try:
+    pkg_resources.get_distribution('plone.app.contenttypes')
+except pkg_resources.DistributionNotFound:
+    from plone.app.testing import PLONE_FIXTURE
+else:
+    from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE as PLONE_FIXTURE
 
 
 class Fixture(PloneSandboxLayer):