+++ CHANGES.rst
@@ -6,7 +6,8 @@ There's a frood who really knows where his towel is.
 2.0b1 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+- Fix upgrade step to work with both, Archetypes and Dexterity-based collections.
+  [rodfersou]
 
 
 2.0a1 (2015-09-30)